requirements.txt
streamlit==1.26.0
opencv-python==4.8.0.74
numpy==1.24.4
ultralytics==8.0.123
Pillow==9.5.0
